Skip the au023 drop of camera kernel
Bug: 143047891

+++ drivers/cam_sensor_module/cam_flash/cam_flash_dev.h
@@ -86,7 +86,7 @@ struct cam_flash_intf_params {
 struct cam_flash_common_attr {
 	bool      is_settings_valid;
 	uint64_t  request_id;
-	uint32_t  count;
+	uint16_t  count;
 	uint8_t   cmd_type;
 };
 
@@ -97,7 +97,7 @@ struct cam_flash_common_attr {
  */
 struct cam_flash_init_packet {
 	struct cam_flash_common_attr  cmn_attr;
-	uint32_t                      flash_type;
+	uint8_t                       flash_type;
 };
 
 /**
